refactor: noquotacard -> smaller files
